refactor: jwt 관련 함수를 utils로 분리
